+++ core-ui/src/main/java/com/anytypeio/anytype/core_ui/features/editor/holders/text/Bulleted.kt
@@ -7,10 +7,7 @@ import com.anytypeio.anytype.core_ui.R
 import com.anytypeio.anytype.core_ui.common.ThemeColor
 import com.anytypeio.anytype.core_ui.extensions.color
 import com.anytypeio.anytype.core_ui.extensions.tint
-import com.anytypeio.anytype.core_ui.features.page.BlockView
-import com.anytypeio.anytype.core_ui.features.page.ListenerType
-import com.anytypeio.anytype.core_ui.features.page.SupportNesting
-import com.anytypeio.anytype.core_ui.features.page.marks
+import com.anytypeio.anytype.core_ui.features.page.*
 import com.anytypeio.anytype.core_ui.widgets.text.TextInputWidget
 import com.anytypeio.anytype.core_utils.ext.dimen
 import kotlinx.android.synthetic.main.item_block_bulleted.view.*
@@ -36,6 +33,7 @@ class Bulleted(
         onSelectionChanged: (String, IntRange) -> Unit,
         onFocusChanged: (String, Boolean) -> Unit,
         clicked: (ListenerType) -> Unit,
+        onMentionEvent: (MentionEvent) -> Unit,
         onSplitLineEnterClicked: (String, Editable, IntRange) -> Unit,
         onEmptyBlockBackspaceClicked: (String) -> Unit,
         onNonEmptyBlockBackspaceClicked: (String, Editable) -> Unit,
@@ -56,7 +54,9 @@ class Bulleted(
         onSplitLineEnterClicked = onSplitLineEnterClicked,
         onNonEmptyBlockBackspaceClicked = onNonEmptyBlockBackspaceClicked,
         onTextInputClicked = onTextInputClicked
-    )
+    ).also {
+        setupMentionWatcher(onMentionEvent)
+    }
 
     override fun getMentionImageSizeAndPadding(): Pair<Int, Int> = with(itemView) {
         Pair(